backend/pyproject.toml
[project]
name = "financial-planner-backend"
version = "0.1.0"
description = "Multi-Agent Financial Analysis System with LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "python-multipart>=0.0.18",

    # Jupyter & Notebook
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "nest-asyncio>=1.6.0",

    # LangChain Core
    "langchain>=1.0.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-ollama>=1.0.0",

    # LangGraph
    "langgraph>=0.3.0",
    "langgraph-cli>=0.4.12",

    # LangSmith for tracing
    "langsmith>=0.2.0",

    # Vector Store
    "qdrant-client>=1.12.0",
    "langchain-qdrant>=0.2.0",

    # OpenAI
    "openai>=1.59.7",

    # Google Sheets
    "gspread>=6.2.1",
    "google-auth>=2.48.0",

    # Data validation
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0",

    # Environment
    "python-dotenv>=1.0.1",

    # Web search
    "tavily-python>=0.7.19",

    # RAG & Evaluations
    "ragas>=0.2.0",
    "beautifulsoup4>=4.12.0",
    "cohere>=5.0.0",
    "langchain-cohere>=0.3.0",
    "requests>=2.31.0",
    "html2text>=2024.2.26",

    # Testing
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
]
